+++ lib/src/components/Minimap.tsx
@@ -1,0 +1,32 @@
+import React from 'react'
+import { observer } from 'mobx-react'
+import { MsaViewModel } from '../model'
+
+const Minimap = observer(function ({ model }: { model: MsaViewModel }) {
+  const { scrollX, msaAreaWidth: W, colWidth, numColumns } = model
+  const unit = W / numColumns
+  const start = -scrollX / colWidth
+  const end = -scrollX / colWidth + W / colWidth
+  const s = start * unit
+  const e = end * unit
+  const HEIGHT = 56
+  const TOP = 10
+  const fill = 'rgba(66, 119, 127, 0.3)'
+  return (
+    <svg height={HEIGHT} style={{ width: '100%' }}>
+      <rect x={0} y={0} width={W} height={TOP} stroke="black" fill="none" />
+      <rect x={s} y={0} width={e - s} height={TOP} fill={fill} />
+      <polygon
+        fill={fill}
+        points={[
+          [e, TOP],
+          [s, TOP],
+          [0, HEIGHT],
+          [W, HEIGHT],
+        ].toString()}
+      />
+    </svg>
+  )
+})
+
+export default Minimap
